Adds a check for wildcard expansion before proceeding with the conversion.

+++ audible-tool.sh
@@ -19,23 +19,25 @@ else
 fi
 
 # Iterate over *.aax files in consume directory and convert to *.m4b if not previously processed.
-for FILE in "${CONSUME_DIR}"/*.aax; do
-  FILENAME=${FILE##*/}
-  if [ ! -f "${CONSUME_DIR}/.processed" ]; then
-    touch "${CONSUME_DIR}/.processed"
-  fi
-
-  if ! grep "${FILENAME}" "${CONSUME_DIR}/.processed" &> /dev/null; then
-    echo "Converting ${FILE}"
-    echo "-> ${STAGING_DIR}/${FILENAME%.*}.m4b"
-    ffmpeg -loglevel error -activation_bytes "$ACTIVATION_BYTES" -i "$FILE" -c copy "${STAGING_DIR}/${FILENAME%.*}.m4b" && \
-    echo "$FILENAME" >> "${CONSUME_DIR}/.processed"
-  else
-    echo "Skipping processed file ${FILE}."
-  fi
-
-  # Cleanup processed file if desired
-  if [ "$REMOVE_PROCESSED_FILES" != "0" ]; then
-    rm "$FILE"
-  fi
-done
+if [ "${CONSUME_DIR}/*.aax" != "$(echo ${CONSUME_DIR}/*.aax)" ]; then
+  for FILE in "${CONSUME_DIR}"/*.aax; do
+    FILENAME=${FILE##*/}
+    if [ ! -f "${CONSUME_DIR}/.processed" ]; then
+      touch "${CONSUME_DIR}/.processed"
+    fi
+  
+    if ! grep "${FILENAME}" "${CONSUME_DIR}/.processed" &> /dev/null; then
+      echo "Converting ${FILE}"
+      echo "-> ${STAGING_DIR}/${FILENAME%.*}.m4b"
+      ffmpeg -loglevel error -activation_bytes "$ACTIVATION_BYTES" -i "$FILE" -c copy "${STAGING_DIR}/${FILENAME%.*}.m4b" && \
+      echo "$FILENAME" >> "${CONSUME_DIR}/.processed"
+    else
+      echo "Skipping processed file ${FILE}."
+    fi
+  
+    # Cleanup processed file if desired
+    if [ "$REMOVE_PROCESSED_FILES" != "0" ]; then
+      rm "$FILE"
+    fi
+  done
+fi
